kg/requirements.txt
playwright
setuptools
wheel
gliner
streamlit
pandas
openai
python-dotenv
neo4j
stvis
pyvis
streamlit-agraph